Fixed task for theme "JCF"
