+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/web_service/operations/ChangeLoginOperation.java
@@ -1,6 +1,6 @@
-package com.rakovets.course.java.core.practice.jcf_set.web_service.Exceptions.operations;
+package com.rakovets.course.java.core.practice.jcf_set.web_service.operations;
 
-import com.rakovets.course.java.core.practice.jcf_set.UsersDataBase;
+import com.rakovets.course.java.core.practice.jcf_set.web_service.UsersDataBase;
 
 public class ChangeLoginOperation extends Operation {
     String login;
