+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/web_service/operations/ChangePasswordOperation.java
@@ -1,6 +1,6 @@
-package com.rakovets.course.java.core.practice.jcf_set.web_service.Exceptions.operations;
+package com.rakovets.course.java.core.practice.jcf_set.web_service.operations;
 
-import com.rakovets.course.java.core.practice.jcf_set.UsersDataBase;
+import com.rakovets.course.java.core.practice.jcf_set.web_service.UsersDataBase;
 
 public class ChangePasswordOperation extends Operation {
     String login;
