+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/web_service/operations/comaparator/OperationsComparator.java
@@ -1,7 +1,6 @@
-package com.rakovets.course.java.core.practice.jcf_set.web_service.Exceptions.operations.comaparator;
+package com.rakovets.course.java.core.practice.jcf_set.web_service.operations.comaparator;
 
-import com.rakovets.course.java.core.practice.jcf_set.web_service.Exceptions.operations.Operation;
-
+import com.rakovets.course.java.core.practice.jcf_set.web_service.operations.Operation;
 import java.util.Comparator;
 
 public class OperationsComparator implements Comparator<Operation> {
